Extract common parts of events testing
